requirements.txt
fastapi==0.115.13
SQLAlchemy==2.0.41
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-decouple==3.8
bcrypt==4.3.0
passlib==1.7.4
itsdangerous==2.2.0
alembic==1.16.2
google-generativeai
uvicorn
python-dotenv